pyproject.toml
[project]
name = "voiceAgent-pipecat"
version = "0.1.0"
description = "Voice Agent with Pipecat and Azure Integration"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.11, <3.13"
dependencies = [
    "aenum==3.1.16",
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "async-timeout==5.0.1",
    "attrs==25.3.0",
    "azure-common==1.1.28",
    "azure-communication-callautomation==1.4.0b1",
    "azure-core==1.35.0",
    "azure-eventgrid==4.22.0",
    "azure-identity==1.23.1",
    "azure-search-documents==11.5.3",
    "cachetools==5.5.2",
    "cartesia==2.0.5",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "coloredlogs==15.0.1",
    "cryptography==45.0.5",
    "dataclasses-json==0.6.7",
    "deepgram-sdk==4.1.1",
    "deprecation==2.1.0",
    "distro==1.9.0",
    "exceptiongroup==1.3.0",
    "fastapi==0.115.12",
    "flatbuffers==25.2.10",
    "frozenlist==1.7.0",
    "future==1.0.0",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.0",
    "google-api-python-client==2.172.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-cloud-speech==2.32.0",
    "google-cloud-texttospeech==2.27.0",
    "google-genai==1.19.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "groq==0.26.0",
    "grpcio==1.73.0",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "humanfriendly==10.0",
    "idna==3.10",
    "isodate==0.7.2",
    "iterators==0.2.0",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain-core==0.3.65",
    "langchain-groq==0.3.2",
    "langsmith==0.3.45",
    "llvmlite==0.44.0",
    "loguru==0.7.3",
    "markdown==3.8",
    "marshmallow==3.26.1",
    "mpmath==1.3.0",
    "msal==1.32.3",
    "msal-extensions==1.3.1",
    "msrest==0.7.1",
    "multidict==6.4.4",
    "mypy-extensions==1.1.0",
    "numba==0.61.2",
    "numpy==1.26.4",
    "oauthlib==3.3.1",
    "onnxruntime==1.20.1",
    "openai==1.70.0",
    "orjson==3.10.18",
    "packaging==24.2",
    "pillow==11.1.0",
    "pipecat-ai==0.0.71",
    "pre-commit>=4.2.0",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydub==0.25.1",
    "pyjwt==2.10.1",
    "pyloudnorm==0.1.1",
    "pyparsing==3.2.3",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "redis==6.2.0",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "resampy==0.4.3",
    "rsa==4.9.1",
    "scipy==1.15.3",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soxr==0.5.0.post1",
    "starlette==0.46.2",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tqdm==4.67.1",
    "typing-extensions==4.14.0",
    "typing-inspect==0.9.0",
    "uritemplate==4.2.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.3",
    "websockets==13.1",
    "yarl==1.20.1",
    "zstandard==0.23.0",
]
